fix: update comments/interfaces for consistency
