+++ protocol/contracts/interfaces/oud/IOudToken.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.8.17;
 // SPDX-License-Identifier: AGPL-3.0-or-later
-// OudToken Interface (protocol/contracts/interfaces/oud/IOudToken.sol)
+// Temple (protocol/contracts/interfaces/oud/IOudToken.sol)
 
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IERC20Permit} from "@openzeppelin/contracts/token/ERC20/extensions/draft-IERC20Permit.sol";
