+++ protocol/contracts/interfaces/templeLineOfCredit/ITempleLineOfCredit.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.8.17;
 // SPDX-License-Identifier: AGPL-3.0-or-later
-// Temple Line of Credit Interface (/interfaces/templeLineOfCredit/ITempleLineOfCredit.sol)
+// Temple (protocol/contracts/interfaces/templeLineOfCredit/ITempleLineOfCredit.sol)
 
 import {IERC20} from '@openzeppelin/contracts/token/ERC20/IERC20.sol';
